common: accept config files containing whitespaces
At the moment, the ceph config file location cannot contain
whitespaces. This applies to all platforms but is quite an
inconvenient on Windows in particular, where it's very common
to have default paths containing white spaces.

ceph_resolve_file_search expects a list of config files and uses
"get_str_list" to separate them. No delimiter list is explicitly
specified, so the default delimiter list is used, which includes
white space.

This change will explicitly set the delimiters to "," and ";".

Signed-off-by: Lucian Petrut <lpetrut@cloudbasesolutions.com>