tests/pybind/cephfs: cleanup xattrs before starting tests
Some xattr tests do not fully cleanup set xattrs. Subsequent
tests may expect xattrs to be absent during the test, such as
setting an xattr and then removing followed by checking if the
xattr list to be empty. This may fail if earlier tests do not
cleanup xattrs, especially for root.

So, cleanup xattrs on root before starting tests. Other directories
are removed anyway, so we do not have to bother about those.

Signed-off-by: Venky Shankar <vshankar@redhat.com>