rpm: build ceph-mgr-diskprediction-local only on certain platforms
d13b441cdb63e37ad3b8ba36cbd3edc33b1e4bbd added a new runtime dependency,
python3-scikit-learn, to the ceph-mgr-diskprediction-local subpackage.

It was later discovered that this dependency is not available on certain
platforms, so 338f47fe948643babd3ae5e640b28ffb1aa1958e was added to
address that.

While 338f47fe948643babd3ae5e640b28ffb1aa1958e takes openSUSE into
account, it does not treat SUSE Linux Enterprise, where the
python3-scikit-learn package is only available via an unsupported repo
(Package Hub).

It was also noted that it doesn't make sense to build/ship a package
that is simply broken (due to missing runtime dependencies), so in
addition to the above, this patch changes the spec file so the package
is built only on those platforms where the dependencies are known to be
available.

Fixes: d13b441cdb63e37ad3b8ba36cbd3edc33b1e4bbd
Signed-off-by: Nathan Cutler <ncutler@suse.com>